requirements.txt
python-telegram-bot==22.2
requests==2.32.3
aiohttp==3.8.5
gunicorn==20.1.0